crypto: use crypto provider to produce ticketer
