+++ rustls/src/ticketer.rs
@@ -1,9 +1,9 @@
-use crate::crypto::{CryptoProvider, Ring};
+#[cfg(test)]
+use crate::crypto::Ring;
+use crate::crypto::CryptoProvider;
 use crate::rand;
 use crate::server::ProducesTickets;
 use crate::Error;
-
-use ring::aead;
 
 use std::mem;
 use std::sync::{Arc, Mutex, MutexGuard};
@@ -27,84 +27,6 @@ impl TimeBase {
     #[inline]
     pub fn as_secs(&self) -> u64 {
         self.0.as_secs()
-    }
-}
-
-/// This is a `ProducesTickets` implementation which uses
-/// any *ring* `aead::Algorithm` to encrypt and authentication
-/// the ticket payload.  It does not enforce any lifetime
-/// constraint.
-struct AeadTicketer {
-    alg: &'static aead::Algorithm,
-    key: aead::LessSafeKey,
-    lifetime: u32,
-}
-
-impl AeadTicketer {
-    /// Make a ticketer with recommended configuration and a random key.
-    fn new() -> Result<Self, rand::GetRandomFailed> {
-        let mut key = [0u8; 32];
-        Ring::fill_random(&mut key)?;
-
-        let alg = &aead::CHACHA20_POLY1305;
-        let key = aead::UnboundKey::new(alg, &key).unwrap();
-
-        Ok(Self {
-            alg,
-            key: aead::LessSafeKey::new(key),
-            lifetime: 60 * 60 * 12,
-        })
-    }
-}
-
-impl ProducesTickets for AeadTicketer {
-    fn enabled(&self) -> bool {
-        true
-    }
-    fn lifetime(&self) -> u32 {
-        self.lifetime
-    }
-
-    /// Encrypt `message` and return the ciphertext.
-    fn encrypt(&self, message: &[u8]) -> Option<Vec<u8>> {
-        // Random nonce, because a counter is a privacy leak.
-        let mut nonce_buf = [0u8; 12];
-        Ring::fill_random(&mut nonce_buf).ok()?;
-        let nonce = ring::aead::Nonce::assume_unique_for_key(nonce_buf);
-        let aad = ring::aead::Aad::empty();
-
-        let mut ciphertext =
-            Vec::with_capacity(nonce_buf.len() + message.len() + self.key.algorithm().tag_len());
-        ciphertext.extend(nonce_buf);
-        ciphertext.extend(message);
-        self.key
-            .seal_in_place_separate_tag(nonce, aad, &mut ciphertext[nonce_buf.len()..])
-            .map(|tag| {
-                ciphertext.extend(tag.as_ref());
-                ciphertext
-            })
-            .ok()
-    }
-
-    /// Decrypt `ciphertext` and recover the original message.
-    fn decrypt(&self, ciphertext: &[u8]) -> Option<Vec<u8>> {
-        // Non-panicking `let (nonce, ciphertext) = ciphertext.split_at(...)`.
-        let nonce = ciphertext.get(..self.alg.nonce_len())?;
-        let ciphertext = ciphertext.get(nonce.len()..)?;
-
-        // This won't fail since `nonce` has the required length.
-        let nonce = ring::aead::Nonce::try_assume_unique_for_key(nonce).ok()?;
-
-        let mut out = Vec::from(ciphertext);
-
-        let plain_len = self
-            .key
-            .open_in_place(nonce, aead::Aad::empty(), &mut out)
-            .ok()?
-            .len();
-        out.truncate(plain_len);
-
-        Some(out)
     }
 }
 
@@ -263,23 +185,22 @@ impl ProducesTickets for TicketSwitcher {
 /// A concrete, safe ticket creation mechanism.
 pub struct Ticketer {}
 
-fn generate_inner() -> Result<Box<dyn ProducesTickets>, rand::GetRandomFailed> {
-    Ok(Box::new(AeadTicketer::new()?))
-}
-
 impl Ticketer {
     /// Make the recommended Ticketer.  This produces tickets
     /// with a 12 hour life and randomly generated keys.
     ///
     /// The encryption mechanism used in Chacha20Poly1305.
-    pub fn new() -> Result<Arc<dyn ProducesTickets>, Error> {
-        Ok(Arc::new(TicketSwitcher::new(6 * 60 * 60, generate_inner)?))
+    pub fn new<C: CryptoProvider>() -> Result<Arc<dyn ProducesTickets>, Error> {
+        Ok(Arc::new(TicketSwitcher::new(
+            6 * 60 * 60,
+            C::ticket_generator,
+        )?))
     }
 }
 
 #[test]
 fn basic_pairwise_test() {
-    let t = Ticketer::new().unwrap();
+    let t = Ticketer::new::<Ring>().unwrap();
     assert!(t.enabled());
     let cipher = t.encrypt(b"hello world").unwrap();
     let plain = t.decrypt(&cipher).unwrap();
@@ -288,7 +209,7 @@ fn basic_pairwise_test() {
 
 #[test]
 fn ticketswitcher_switching_test() {
-    let t = Arc::new(TicketSwitcher::new(1, generate_inner).unwrap());
+    let t = Arc::new(TicketSwitcher::new(1, Ring::ticket_generator).unwrap());
     let now = TimeBase::now().unwrap();
     let cipher1 = t.encrypt(b"ticket 1").unwrap();
     assert_eq!(t.decrypt(&cipher1).unwrap(), b"ticket 1");
@@ -316,7 +237,7 @@ fn fail_generator() -> Result<Box<dyn ProducesTickets>, rand::GetRandomFailed> {
 
 #[test]
 fn ticketswitcher_recover_test() {
-    let mut t = TicketSwitcher::new(1, generate_inner).unwrap();
+    let mut t = TicketSwitcher::new(1, Ring::ticket_generator).unwrap();
     let now = TimeBase::now().unwrap();
     let cipher1 = t.encrypt(b"ticket 1").unwrap();
     assert_eq!(t.decrypt(&cipher1).unwrap(), b"ticket 1");
@@ -325,7 +246,7 @@ fn ticketswitcher_recover_test() {
         // Failed new ticketer
         t.maybe_roll(TimeBase(now.0 + std::time::Duration::from_secs(10)));
     }
-    t.generator = generate_inner;
+    t.generator = Ring::ticket_generator;
     let cipher2 = t.encrypt(b"ticket 2").unwrap();
     assert_eq!(t.decrypt(&cipher1).unwrap(), b"ticket 1");
     assert_eq!(t.decrypt(&cipher2).unwrap(), b"ticket 2");
